refactor: Certificacao
Realocar classe Certificacao p/ raiz da Models
