+++ backend/src/main/java/com/senac/gestaocurso/repository/CertifcacaoRepository.java
@@ -1,7 +1,7 @@
 package com.senac.gestaocurso.repository;
 
 import com.senac.gestaocurso.enterprise.CustomQuerydslPredicateExecutor;
-import com.senac.gestaocurso.models.domain.Certificacao;
+import com.senac.gestaocurso.models.Certificacao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
